+++ FreeRTOS_TCP_IP.c
@@ -59,106 +59,106 @@
 
 /*lint -e750  local macro not referenced [MISRA 2012 Rule 2.5, advisory] */
 
-/*
- * The meaning of the TCP flags:
- */
-    #define tcpTCP_FLAG_FIN     ( ( uint8_t ) 0x01U )                   /**< No more data from sender. */
-    #define tcpTCP_FLAG_SYN     ( ( uint8_t ) 0x02U )                   /**< Synchronize sequence numbers. */
-    #define tcpTCP_FLAG_RST     ( ( uint8_t ) 0x04U )                   /**< Reset the connection. */
-    #define tcpTCP_FLAG_PSH     ( ( uint8_t ) 0x08U )                   /**< Push function: please push buffered data to the recv application. */
-    #define tcpTCP_FLAG_ACK     ( ( uint8_t ) 0x10U )                   /**< Acknowledgment field is significant. */
-    #define tcpTCP_FLAG_URG     ( ( uint8_t ) 0x20U )                   /**< Urgent pointer field is significant. */
-    #define tcpTCP_FLAG_ECN     ( ( uint8_t ) 0x40U )                   /**< ECN-Echo. */
-    #define tcpTCP_FLAG_CWR     ( ( uint8_t ) 0x80U )                   /**< Congestion Window Reduced. */
+// /*
+//  * The meaning of the TCP flags:
+//  */
+//     #define tcpTCP_FLAG_FIN     ( ( uint8_t ) 0x01U )                   /**< No more data from sender. */
+//     #define tcpTCP_FLAG_SYN     ( ( uint8_t ) 0x02U )                   /**< Synchronize sequence numbers. */
+//     #define tcpTCP_FLAG_RST     ( ( uint8_t ) 0x04U )                   /**< Reset the connection. */
+//     #define tcpTCP_FLAG_PSH     ( ( uint8_t ) 0x08U )                   /**< Push function: please push buffered data to the recv application. */
+//     #define tcpTCP_FLAG_ACK     ( ( uint8_t ) 0x10U )                   /**< Acknowledgment field is significant. */
+//     #define tcpTCP_FLAG_URG     ( ( uint8_t ) 0x20U )                   /**< Urgent pointer field is significant. */
+//     #define tcpTCP_FLAG_ECN     ( ( uint8_t ) 0x40U )                   /**< ECN-Echo. */
+//     #define tcpTCP_FLAG_CWR     ( ( uint8_t ) 0x80U )                   /**< Congestion Window Reduced. */
 
-    #define tcpTCP_FLAG_CTRL    ( ( uint8_t ) 0x1FU )                   /**< A mask to filter all protocol flags. */
-
-
-/*
- * A few values of the TCP options:
- */
-    #define tcpTCP_OPT_END               0U          /**< End of TCP options list. */
-    #define tcpTCP_OPT_NOOP              1U          /**< "No-operation" TCP option. */
-    #define tcpTCP_OPT_MSS               2U          /**< Maximum segment size TCP option. */
-    #define tcpTCP_OPT_WSOPT             3U          /**< TCP Window Scale Option (3-byte long). */
-    #define tcpTCP_OPT_SACK_P            4U          /**< Advertise that SACK is permitted. */
-    #define tcpTCP_OPT_SACK_A            5U          /**< SACK option with first/last. */
-    #define tcpTCP_OPT_TIMESTAMP         8U          /**< Time-stamp option. */
+//     #define tcpTCP_FLAG_CTRL    ( ( uint8_t ) 0x1FU )                   /**< A mask to filter all protocol flags. */
 
 
-    #define tcpTCP_OPT_MSS_LEN           4U          /**< Length of TCP MSS option. */
-    #define tcpTCP_OPT_WSOPT_LEN         3U          /**< Length of TCP WSOPT option. */
-
-    #define tcpTCP_OPT_TIMESTAMP_LEN     10          /**< fixed length of the time-stamp option. */
-
-/** @brief
- * Minimum segment length as outlined by RFC 791 section 3.1.
- * Minimum segment length ( 536 ) = Minimum MTU ( 576 ) - IP Header ( 20 ) - TCP Header ( 20 ).
- */
-    #define tcpMINIMUM_SEGMENT_LENGTH    536U
-
-/** @brief
- * The macro tcpNOW_CONNECTED() is use to determine if the connection makes a
- * transition from connected to non-connected and vice versa.
- * tcpNOW_CONNECTED() returns true when the status has one of these values:
- * eESTABLISHED, eFIN_WAIT_1, eFIN_WAIT_2, eCLOSING, eLAST_ACK, eTIME_WAIT
- * Technically the connection status is closed earlier, but the library wants
- * to prevent that the socket will be deleted before the last ACK has been
- * and thus causing a 'RST' packet on either side.
- */
-    #define tcpNOW_CONNECTED( status ) \
-    ( ( ( ( status ) >= ( BaseType_t ) eESTABLISHED ) && ( ( status ) != ( BaseType_t ) eCLOSE_WAIT ) ) ? 1 : 0 )
-
-/** @brief
- * The highest 4 bits in the TCP offset byte indicate the total length of the
- * TCP header, divided by 4.
- */
-    #define tcpVALID_BITS_IN_TCP_OFFSET_BYTE    ( 0xF0U )
-
-/*
- * Acknowledgements to TCP data packets may be delayed as long as more is being expected.
- * A normal delay would be 200ms. Here a much shorter delay of 20 ms is being used to
- * gain performance.
- */
-    #define tcpDELAYED_ACK_SHORT_DELAY_MS       ( 2 )   /**< Should not become smaller than 1. */
-    #define tcpDELAYED_ACK_LONGER_DELAY_MS      ( 20 )  /**< Longer delay for ACK. */
+// /*
+//  * A few values of the TCP options:
+//  */
+//     #define tcpTCP_OPT_END               0U          /**< End of TCP options list. */
+//     #define tcpTCP_OPT_NOOP              1U          /**< "No-operation" TCP option. */
+//     #define tcpTCP_OPT_MSS               2U          /**< Maximum segment size TCP option. */
+//     #define tcpTCP_OPT_WSOPT             3U          /**< TCP Window Scale Option (3-byte long). */
+//     #define tcpTCP_OPT_SACK_P            4U          /**< Advertise that SACK is permitted. */
+//     #define tcpTCP_OPT_SACK_A            5U          /**< SACK option with first/last. */
+//     #define tcpTCP_OPT_TIMESTAMP         8U          /**< Time-stamp option. */
 
 
-/** @brief
- * The MSS (Maximum Segment Size) will be taken as large as possible. However, packets with
- * an MSS of 1460 bytes won't be transported through the internet.  The MSS will be reduced
- * to 1400 bytes.
- */
-    #define tcpREDUCED_MSS_THROUGH_INTERNET    ( 1400 )
+//     #define tcpTCP_OPT_MSS_LEN           4U          /**< Length of TCP MSS option. */
+//     #define tcpTCP_OPT_WSOPT_LEN         3U          /**< Length of TCP WSOPT option. */
 
-/** @brief
- * When there are no TCP options, the TCP offset equals 20 bytes, which is stored as
- * the number 5 (words) in the higher nibble of the TCP-offset byte.
- */
-    #define tcpTCP_OFFSET_LENGTH_BITS          ( 0xf0U )
-    #define tcpTCP_OFFSET_STANDARD_LENGTH      ( 0x50U )  /**< Standard TCP packet offset. */
+//     #define tcpTCP_OPT_TIMESTAMP_LEN     10          /**< fixed length of the time-stamp option. */
+
+// /** @brief
+//  * Minimum segment length as outlined by RFC 791 section 3.1.
+//  * Minimum segment length ( 536 ) = Minimum MTU ( 576 ) - IP Header ( 20 ) - TCP Header ( 20 ).
+//  */
+//     #define tcpMINIMUM_SEGMENT_LENGTH    536U
+
+// /** @brief
+//  * The macro tcpNOW_CONNECTED() is use to determine if the connection makes a
+//  * transition from connected to non-connected and vice versa.
+//  * tcpNOW_CONNECTED() returns true when the status has one of these values:
+//  * eESTABLISHED, eFIN_WAIT_1, eFIN_WAIT_2, eCLOSING, eLAST_ACK, eTIME_WAIT
+//  * Technically the connection status is closed earlier, but the library wants
+//  * to prevent that the socket will be deleted before the last ACK has been
+//  * and thus causing a 'RST' packet on either side.
+//  */
+//     #define tcpNOW_CONNECTED( status ) \
+//     ( ( ( ( status ) >= ( BaseType_t ) eESTABLISHED ) && ( ( status ) != ( BaseType_t ) eCLOSE_WAIT ) ) ? 1 : 0 )
+
+// /** @brief
+//  * The highest 4 bits in the TCP offset byte indicate the total length of the
+//  * TCP header, divided by 4.
+//  */
+//     #define tcpVALID_BITS_IN_TCP_OFFSET_BYTE    ( 0xF0U )
+
+// /*
+//  * Acknowledgements to TCP data packets may be delayed as long as more is being expected.
+//  * A normal delay would be 200ms. Here a much shorter delay of 20 ms is being used to
+//  * gain performance.
+//  */
+//     #define tcpDELAYED_ACK_SHORT_DELAY_MS       ( 2 )   /**< Should not become smaller than 1. */
+//     #define tcpDELAYED_ACK_LONGER_DELAY_MS      ( 20 )  /**< Longer delay for ACK. */
 
 
-/** @brief
- * Each TCP socket is checked regularly to see if it can send data packets.
- * By default, the maximum number of packets sent during one check is limited to 8.
- * This amount may be further limited by setting the socket's TX window size.
- */
-    #if ( !defined( SEND_REPEATED_COUNT ) )
-        #define SEND_REPEATED_COUNT    ( 8 )
-    #endif /* !defined( SEND_REPEATED_COUNT ) */
+// /** @brief
+//  * The MSS (Maximum Segment Size) will be taken as large as possible. However, packets with
+//  * an MSS of 1460 bytes won't be transported through the internet.  The MSS will be reduced
+//  * to 1400 bytes.
+//  */
+//     #define tcpREDUCED_MSS_THROUGH_INTERNET    ( 1400 )
 
-/** @brief
- * Define a maximum period of time (ms) to leave a TCP-socket unattended.
- * When a TCP timer expires, retries and keep-alive messages will be checked.
- */
-    #ifndef tcpMAXIMUM_TCP_WAKEUP_TIME_MS
-        #define tcpMAXIMUM_TCP_WAKEUP_TIME_MS    20000U
-    #endif
+// /** @brief
+//  * When there are no TCP options, the TCP offset equals 20 bytes, which is stored as
+//  * the number 5 (words) in the higher nibble of the TCP-offset byte.
+//  */
+//     #define tcpTCP_OFFSET_LENGTH_BITS          ( 0xf0U )
+//     #define tcpTCP_OFFSET_STANDARD_LENGTH      ( 0x50U )  /**< Standard TCP packet offset. */
 
-/* Two macro's that were introduced to work with both IPv4 and IPv6. */
-    #define xIPHeaderSize( pxNetworkBuffer )    ( ipSIZE_OF_IPv4_HEADER )  /**< Size of IP Header. */
-    #define uxIPHeaderSizeSocket( pxSocket )    ( ipSIZE_OF_IPv4_HEADER )  /**< Size of IP Header socket. */
+
+// /** @brief
+//  * Each TCP socket is checked regularly to see if it can send data packets.
+//  * By default, the maximum number of packets sent during one check is limited to 8.
+//  * This amount may be further limited by setting the socket's TX window size.
+//  */
+//     #if ( !defined( SEND_REPEATED_COUNT ) )
+//         #define SEND_REPEATED_COUNT    ( 8 )
+//     #endif /* !defined( SEND_REPEATED_COUNT ) */
+
+// /** @brief
+//  * Define a maximum period of time (ms) to leave a TCP-socket unattended.
+//  * When a TCP timer expires, retries and keep-alive messages will be checked.
+//  */
+//     #ifndef tcpMAXIMUM_TCP_WAKEUP_TIME_MS
+//         #define tcpMAXIMUM_TCP_WAKEUP_TIME_MS    20000U
+//     #endif
+
+// /* Two macro's that were introduced to work with both IPv4 and IPv6. */
+//     #define xIPHeaderSize( pxNetworkBuffer )    ( ipSIZE_OF_IPv4_HEADER )  /**< Size of IP Header. */
+//     #define uxIPHeaderSizeSocket( pxSocket )    ( ipSIZE_OF_IPv4_HEADER )  /**< Size of IP Header socket. */
 
 
 // /*
@@ -197,7 +197,7 @@
 //  */
 //     static BaseType_t prvTCPPrepareConnect( FreeRTOS_Socket_t * pxSocket );
 
-    #if ( ipconfigHAS_DEBUG_PRINTF != 0 )
+//    #if ( ipconfigHAS_DEBUG_PRINTF != 0 )
 
 // /*
 //  * For logging and debugging: make a string showing the TCP flags.
@@ -810,1196 +810,6 @@
     /*-----------------------------------------------------------*/
 
 /**
- * @brief The API FreeRTOS_send() adds data to the TX stream. Add
- *        this data to the windowing system to it can be transmitted.
- *
- * @param[in] pxSocket: The socket owning the connection.
- */
-    static void prvTCPAddTxData( FreeRTOS_Socket_t * pxSocket )
-    {
-        int32_t lCount, lLength;
-
-        /* A txStream has been created already, see if the socket has new data for
-         * the sliding window.
-         *
-         * uxStreamBufferMidSpace() returns the distance between rxHead and rxMid.  It
-         * contains new Tx data which has not been passed to the sliding window yet.
-         * The oldest data not-yet-confirmed can be found at rxTail. */
-        lLength = ( int32_t ) uxStreamBufferMidSpace( pxSocket->u.xTCP.txStream );
-
-        if( lLength > 0 )
-        {
-            /* All data between txMid and rxHead will now be passed to the sliding
-             * window manager, so it can start transmitting them.
-             *
-             * Hand over the new data to the sliding window handler.  It will be
-             * split-up in chunks of 1460 bytes each (or less, depending on
-             * ipconfigTCP_MSS). */
-            lCount = lTCPWindowTxAdd( &pxSocket->u.xTCP.xTCPWindow,
-                                      ( uint32_t ) lLength,
-                                      ( int32_t ) pxSocket->u.xTCP.txStream->uxMid,
-                                      ( int32_t ) pxSocket->u.xTCP.txStream->LENGTH );
-
-            /* Move the rxMid pointer forward up to rxHead. */
-            if( lCount > 0 )
-            {
-                vStreamBufferMoveMid( pxSocket->u.xTCP.txStream, ( size_t ) lCount );
-            }
-        }
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief prvTCPHandleFin() will be called to handle connection closure. The
- *        closure starts when either a FIN has been received and accepted,
- *        or when the socket has sent a FIN flag to the peer. Before being
- *        called, it has been checked that both reception and transmission
- *        are complete.
- *
- * @param[in] pxSocket: Socket owning the the connection.
- * @param[in] pxNetworkBuffer: The network buffer carrying the TCP packet.
- *
- * @return Length of the packet to be sent.
- */
-    static BaseType_t prvTCPHandleFin( FreeRTOS_Socket_t * pxSocket,
-                                       const NetworkBufferDescriptor_t * pxNetworkBuffer )
-    {
-        /* Map the ethernet buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        ProtocolHeaders_t * pxProtocolHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                        &( pxNetworkBuffer->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + xIPHeaderSize( pxNetworkBuffer ) ] ) );
-        TCPHeader_t * pxTCPHeader = &( pxProtocolHeaders->xTCPHeader );
-        uint8_t ucIntermediateResult = 0, ucTCPFlags = pxTCPHeader->ucTCPFlags;
-        TCPWindow_t * pxTCPWindow = &pxSocket->u.xTCP.xTCPWindow;
-        BaseType_t xSendLength = 0;
-        uint32_t ulAckNr = FreeRTOS_ntohl( pxTCPHeader->ulAckNr );
-
-        if( ( ucTCPFlags & tcpTCP_FLAG_FIN ) != 0U )
-        {
-            pxTCPWindow->rx.ulCurrentSequenceNumber = pxTCPWindow->rx.ulFINSequenceNumber + 1U;
-        }
-
-        if( pxSocket->u.xTCP.bits.bFinSent == pdFALSE_UNSIGNED )
-        {
-            /* We haven't yet replied with a FIN, do so now. */
-            pxTCPWindow->tx.ulFINSequenceNumber = pxTCPWindow->tx.ulCurrentSequenceNumber;
-            pxSocket->u.xTCP.bits.bFinSent = pdTRUE_UNSIGNED;
-        }
-        else
-        {
-            /* We did send a FIN already, see if it's ACK'd. */
-            if( ulAckNr == ( pxTCPWindow->tx.ulFINSequenceNumber + 1U ) )
-            {
-                pxSocket->u.xTCP.bits.bFinAcked = pdTRUE_UNSIGNED;
-            }
-        }
-
-        if( pxSocket->u.xTCP.bits.bFinAcked == pdFALSE_UNSIGNED )
-        {
-            pxTCPWindow->tx.ulCurrentSequenceNumber = pxTCPWindow->tx.ulFINSequenceNumber;
-            pxTCPHeader->ucTCPFlags = ( uint8_t ) tcpTCP_FLAG_ACK | ( uint8_t ) tcpTCP_FLAG_FIN;
-
-            /* And wait for the final ACK. */
-            vTCPStateChange( pxSocket, eLAST_ACK );
-        }
-        else
-        {
-            /* Our FIN has been ACK'd, the outgoing sequence number is now fixed. */
-            pxTCPWindow->tx.ulCurrentSequenceNumber = pxTCPWindow->tx.ulFINSequenceNumber + 1U;
-
-            if( pxSocket->u.xTCP.bits.bFinRecv == pdFALSE_UNSIGNED )
-            {
-                /* We have sent out a FIN but the peer hasn't replied with a FIN
-                 * yet. Do nothing for the moment. */
-                pxTCPHeader->ucTCPFlags = 0U;
-            }
-            else
-            {
-                if( pxSocket->u.xTCP.bits.bFinLast == pdFALSE_UNSIGNED )
-                {
-                    /* This is the third of the three-way hand shake: the last
-                     * ACK. */
-                    pxTCPHeader->ucTCPFlags = tcpTCP_FLAG_ACK;
-                }
-                else
-                {
-                    /* The other party started the closure, so we just wait for the
-                     * last ACK. */
-                    pxTCPHeader->ucTCPFlags = 0U;
-                }
-
-                /* And wait for the user to close this socket. */
-                vTCPStateChange( pxSocket, eCLOSE_WAIT );
-            }
-        }
-
-        pxTCPWindow->ulOurSequenceNumber = pxTCPWindow->tx.ulCurrentSequenceNumber;
-
-        if( pxTCPHeader->ucTCPFlags != 0U )
-        {
-            ucIntermediateResult = uxIPHeaderSizeSocket( pxSocket ) + ipSIZE_OF_TCP_HEADER + pxTCPWindow->ucOptionLength;
-            xSendLength = ( BaseType_t ) ucIntermediateResult;
-        }
-
-        pxTCPHeader->ucTCPOffset = ( uint8_t ) ( ( ipSIZE_OF_TCP_HEADER + pxTCPWindow->ucOptionLength ) << 2 );
-
-        if( xTCPWindowLoggingLevel != 0 )
-        {
-            FreeRTOS_debug_printf( ( "TCP: send FIN+ACK (ack %u, cur/nxt %u/%u) ourSeqNr %u | Rx %u\n",
-                                     ( unsigned ) ( ulAckNr - pxTCPWindow->tx.ulFirstSequenceNumber ),
-                                     ( unsigned ) ( pxTCPWindow->tx.ulCurrentSequenceNumber - pxTCPWindow->tx.ulFirstSequenceNumber ),
-                                     ( unsigned ) ( pxTCPWindow->ulNextTxSequenceNumber - pxTCPWindow->tx.ulFirstSequenceNumber ),
-                                     ( unsigned ) ( pxTCPWindow->ulOurSequenceNumber - pxTCPWindow->tx.ulFirstSequenceNumber ),
-                                     ( unsigned ) ( pxTCPWindow->rx.ulCurrentSequenceNumber - pxTCPWindow->rx.ulFirstSequenceNumber ) ) );
-        }
-
-        return xSendLength;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief prvCheckRxData(): called from prvTCPHandleState(). The
- *        first thing that will be done is find the TCP payload data
- *        and check the length of this data.
- *
- * @param[in] pxNetworkBuffer: The network buffer holding the received data.
- * @param[out] ppucRecvData: It will point to first byte of the TCP payload.
- *
- * @return Length of the received buffer.
- */
-    static BaseType_t prvCheckRxData( const NetworkBufferDescriptor_t * pxNetworkBuffer,
-                                      uint8_t ** ppucRecvData )
-    {
-        /* Map the ethernet buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        const ProtocolHeaders_t * pxProtocolHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                              &( pxNetworkBuffer->pucEthernetBuffer[ ( size_t ) ipSIZE_OF_ETH_HEADER + xIPHeaderSize( pxNetworkBuffer ) ] ) );
-        const TCPHeader_t * pxTCPHeader = &( pxProtocolHeaders->xTCPHeader );
-        int32_t lLength, lTCPHeaderLength, lReceiveLength, lUrgentLength;
-
-        /* Map the buffer onto an IPHeader_t struct for easy access to fields. */
-        const IPHeader_t * pxIPHeader = ipCAST_CONST_PTR_TO_CONST_TYPE_PTR( IPHeader_t, &( pxNetworkBuffer->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER ] ) );
-        const size_t xIPHeaderLength = ipSIZE_OF_IPv4_HEADER;
-        uint16_t usLength;
-        uint8_t ucIntermediateResult = 0;
-
-        /* Determine the length and the offset of the user-data sent to this
-         * node.
-         *
-         * The size of the TCP header is given in a multiple of 4-byte words (single
-         * byte, needs no ntoh() translation).  A shift-right 2: is the same as
-         * (offset >> 4) * 4. */
-        ucIntermediateResult = ( pxTCPHeader->ucTCPOffset & tcpVALID_BITS_IN_TCP_OFFSET_BYTE ) >> 2;
-        lTCPHeaderLength = ( int32_t ) ucIntermediateResult;
-
-        /* Let pucRecvData point to the first byte received. */
-        *ppucRecvData = &( pxNetworkBuffer->pucEthernetBuffer[ ( size_t ) ipSIZE_OF_ETH_HEADER + xIPHeaderLength + ( size_t ) lTCPHeaderLength ] );
-
-        /* Calculate lReceiveLength - the length of the TCP data received.  This is
-         * equal to the total packet length minus:
-         * ( LinkLayer length (14) + IP header length (20) + size of TCP header(20 +) ).*/
-        lReceiveLength = ( int32_t ) pxNetworkBuffer->xDataLength;
-        lReceiveLength -= ( int32_t ) ipSIZE_OF_ETH_HEADER;
-
-        usLength = FreeRTOS_htons( pxIPHeader->usLength );
-        lLength = ( int32_t ) usLength;
-
-        if( lReceiveLength > lLength )
-        {
-            /* More bytes were received than the reported length, often because of
-             * padding bytes at the end. */
-            lReceiveLength = lLength;
-        }
-
-        /* Subtract the size of the TCP and IP headers and the actual data size is
-         * known. */
-        if( lReceiveLength > ( lTCPHeaderLength + ( int32_t ) xIPHeaderLength ) )
-        {
-            lReceiveLength -= ( lTCPHeaderLength + ( int32_t ) xIPHeaderLength );
-        }
-        else
-        {
-            lReceiveLength = 0;
-        }
-
-        /* Urgent Pointer:
-         * This field communicates the current value of the urgent pointer as a
-         * positive offset from the sequence number in this segment.  The urgent
-         * pointer points to the sequence number of the octet following the urgent
-         * data.  This field is only be interpreted in segments with the URG control
-         * bit set. */
-        if( ( pxTCPHeader->ucTCPFlags & tcpTCP_FLAG_URG ) != 0U )
-        {
-            /* Although we ignore the urgent data, we have to skip it. */
-            lUrgentLength = ( int32_t ) FreeRTOS_htons( pxTCPHeader->usUrgent );
-            *ppucRecvData += lUrgentLength;
-            lReceiveLength -= FreeRTOS_min_int32( lReceiveLength, lUrgentLength );
-        }
-
-        return ( BaseType_t ) lReceiveLength;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief prvStoreRxData(): called from prvTCPHandleState().
- *        The second thing is to do is check if the payload data may
- *        be accepted. If so, they will be added to the reception queue.
- *
- * @param[in] pxSocket: The socket owning the connection.
- * @param[in] pucRecvData: Pointer to received data.
- * @param[in] pxNetworkBuffer: The network buffer descriptor.
- * @param[in] ulReceiveLength: The length of the received data.
- *
- * @return 0 on success, -1 on failure of storing data.
- */
-    static BaseType_t prvStoreRxData( FreeRTOS_Socket_t * pxSocket,
-                                      const uint8_t * pucRecvData,
-                                      NetworkBufferDescriptor_t * pxNetworkBuffer,
-                                      uint32_t ulReceiveLength )
-    {
-        /* Map the ethernet buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        const ProtocolHeaders_t * pxProtocolHeaders = ipCAST_CONST_PTR_TO_CONST_TYPE_PTR( ProtocolHeaders_t,
-                                                                                          &( pxNetworkBuffer->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + xIPHeaderSize( pxNetworkBuffer ) ] ) );
-        const TCPHeader_t * pxTCPHeader = &pxProtocolHeaders->xTCPHeader;
-        TCPWindow_t * pxTCPWindow = &pxSocket->u.xTCP.xTCPWindow;
-        uint32_t ulSequenceNumber, ulSpace;
-        int32_t lOffset, lStored;
-        BaseType_t xResult = 0;
-
-        ulSequenceNumber = FreeRTOS_ntohl( pxTCPHeader->ulSequenceNumber );
-
-        if( ( ulReceiveLength > 0U ) && ( pxSocket->u.xTCP.ucTCPState >= ( uint8_t ) eSYN_RECEIVED ) )
-        {
-            /* See if way may accept the data contents and forward it to the socket
-             * owner.
-             *
-             * If it can't be "accept"ed it may have to be stored and send a selective
-             * ack (SACK) option to confirm it.  In that case, lTCPAddRxdata() will be
-             * called later to store an out-of-order packet (in case lOffset is
-             * negative). */
-            if( pxSocket->u.xTCP.rxStream != NULL )
-            {
-                ulSpace = ( uint32_t ) uxStreamBufferGetSpace( pxSocket->u.xTCP.rxStream );
-            }
-            else
-            {
-                ulSpace = ( uint32_t ) pxSocket->u.xTCP.uxRxStreamSize;
-            }
-
-            lOffset = lTCPWindowRxCheck( pxTCPWindow, ulSequenceNumber, ulReceiveLength, ulSpace );
-
-            if( lOffset >= 0 )
-            {
-                /* New data has arrived and may be made available to the user.  See
-                 * if the head marker in rxStream may be advanced, only if lOffset == 0.
-                 * In case the low-water mark is reached, bLowWater will be set
-                 * "low-water" here stands for "little space". */
-                lStored = lTCPAddRxdata( pxSocket, ( uint32_t ) lOffset, pucRecvData, ulReceiveLength );
-
-                if( lStored != ( int32_t ) ulReceiveLength )
-                {
-                    FreeRTOS_debug_printf( ( "lTCPAddRxdata: stored %d / %u bytes? ?\n", ( int ) lStored, ( unsigned ) ulReceiveLength ) );
-
-                    /* Received data could not be stored.  The socket's flag
-                     * bMallocError has been set.  The socket now has the status
-                     * eCLOSE_WAIT and a RST packet will be sent back. */
-                    ( void ) prvTCPSendReset( pxNetworkBuffer );
-                    xResult = -1;
-                }
-            }
-
-            /* After a missing packet has come in, higher packets may be passed to
-             * the user. */
-            #if ( ipconfigUSE_TCP_WIN == 1 )
-                {
-                    /* Now lTCPAddRxdata() will move the rxHead pointer forward
-                     * so data becomes available to the user immediately
-                     * In case the low-water mark is reached, bLowWater will be set. */
-                    if( ( xResult == 0 ) && ( pxTCPWindow->ulUserDataLength > 0U ) )
-                    {
-                        ( void ) lTCPAddRxdata( pxSocket, 0U, NULL, pxTCPWindow->ulUserDataLength );
-                        pxTCPWindow->ulUserDataLength = 0;
-                    }
-                }
-            #endif /* ipconfigUSE_TCP_WIN */
-        }
-        else
-        {
-            pxTCPWindow->ucOptionLength = 0U;
-        }
-
-        return xResult;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief Set the TCP options (if any) for the outgoing packet.
- *
- * @param[in] pxSocket: The socket owning the connection.
- * @param[in] pxNetworkBuffer: The network buffer holding the packet.
- *
- * @return Length of the TCP options after they are set.
- */
-    static UBaseType_t prvSetOptions( FreeRTOS_Socket_t * pxSocket,
-                                      const NetworkBufferDescriptor_t * pxNetworkBuffer )
-    {
-        /* Map the ethernet buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        ProtocolHeaders_t * pxProtocolHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                        &( pxNetworkBuffer->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + xIPHeaderSize( pxNetworkBuffer ) ] ) );
-        TCPHeader_t * pxTCPHeader = &pxProtocolHeaders->xTCPHeader;
-        const TCPWindow_t * pxTCPWindow = &pxSocket->u.xTCP.xTCPWindow;
-        UBaseType_t uxOptionsLength = pxTCPWindow->ucOptionLength;
-
-        #if ( ipconfigUSE_TCP_WIN == 1 )
-            /* memcpy() helper variables for MISRA Rule 21.15 compliance*/
-            const void * pvCopySource;
-            void * pvCopyDest;
-
-            if( uxOptionsLength != 0U )
-            {
-                /* TCP options must be sent because a packet which is out-of-order
-                 * was received. */
-                if( xTCPWindowLoggingLevel >= 0 )
-                {
-                    FreeRTOS_debug_printf( ( "SACK[%u,%u]: optlen %u sending %u - %u\n",
-                                             pxSocket->usLocalPort,
-                                             pxSocket->u.xTCP.usRemotePort,
-                                             ( unsigned ) uxOptionsLength,
-                                             ( unsigned ) ( FreeRTOS_ntohl( pxTCPWindow->ulOptionsData[ 1 ] ) - pxSocket->u.xTCP.xTCPWindow.rx.ulFirstSequenceNumber ),
-                                             ( unsigned ) ( FreeRTOS_ntohl( pxTCPWindow->ulOptionsData[ 2 ] ) - pxSocket->u.xTCP.xTCPWindow.rx.ulFirstSequenceNumber ) ) );
-                }
-
-                /*
-                 * Use helper variables for memcpy() source & dest to remain
-                 * compliant with MISRA Rule 21.15.  These should be
-                 * optimized away.
-                 */
-                pvCopySource = pxTCPWindow->ulOptionsData;
-                pvCopyDest = pxTCPHeader->ucOptdata;
-                ( void ) memcpy( pvCopyDest, pvCopySource, ( size_t ) uxOptionsLength );
-
-                /* The header length divided by 4, goes into the higher nibble,
-                 * effectively a shift-left 2. */
-                pxTCPHeader->ucTCPOffset = ( uint8_t ) ( ( ipSIZE_OF_TCP_HEADER + uxOptionsLength ) << 2 );
-            }
-            else
-        #endif /* ipconfigUSE_TCP_WIN */
-
-        if( ( pxSocket->u.xTCP.ucTCPState >= ( EventBits_t ) eESTABLISHED ) && ( pxSocket->u.xTCP.bits.bMssChange != pdFALSE_UNSIGNED ) )
-        {
-            /* TCP options must be sent because the MSS has changed. */
-            pxSocket->u.xTCP.bits.bMssChange = pdFALSE_UNSIGNED;
-
-            if( xTCPWindowLoggingLevel >= 0 )
-            {
-                FreeRTOS_debug_printf( ( "MSS: sending %u\n", pxSocket->u.xTCP.usMSS ) );
-            }
-
-            pxTCPHeader->ucOptdata[ 0 ] = tcpTCP_OPT_MSS;
-            pxTCPHeader->ucOptdata[ 1 ] = tcpTCP_OPT_MSS_LEN;
-            pxTCPHeader->ucOptdata[ 2 ] = ( uint8_t ) ( ( pxSocket->u.xTCP.usMSS ) >> 8 );
-            pxTCPHeader->ucOptdata[ 3 ] = ( uint8_t ) ( ( pxSocket->u.xTCP.usMSS ) & 0xffU );
-            uxOptionsLength = 4U;
-            pxTCPHeader->ucTCPOffset = ( uint8_t ) ( ( ipSIZE_OF_TCP_HEADER + uxOptionsLength ) << 2 );
-        }
-        else
-        {
-            /* Nothing. */
-        }
-
-        return uxOptionsLength;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief prvHandleSynReceived(): called from prvTCPHandleState(). Called
- *        from the states: eSYN_RECEIVED and eCONNECT_SYN. If the flags
- *        received are correct, the socket will move to eESTABLISHED.
- *
- * @param[in] pxSocket: The socket handling the connection.
- * @param[in] pxNetworkBuffer: The pointer to the network buffer carrying
- *                             the packet.
- * @param[in] ulReceiveLength: Length in bytes of the data received.
- * @param[in] uxOptionsLength: Length of the TCP options in bytes.
- *
- * @return Length of the data to be sent.
- */
-    static BaseType_t prvHandleSynReceived( FreeRTOS_Socket_t * pxSocket,
-                                            const NetworkBufferDescriptor_t * pxNetworkBuffer,
-                                            uint32_t ulReceiveLength,
-                                            UBaseType_t uxOptionsLength )
-    {
-        /* Map the ethernet buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        ProtocolHeaders_t * pxProtocolHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                        &( pxNetworkBuffer->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + uxIPHeaderSizeSocket( pxSocket ) ] ) );
-        TCPHeader_t * pxTCPHeader = &pxProtocolHeaders->xTCPHeader;
-        TCPWindow_t * pxTCPWindow = &pxSocket->u.xTCP.xTCPWindow;
-        uint8_t ucTCPFlags = pxTCPHeader->ucTCPFlags;
-        uint32_t ulSequenceNumber = FreeRTOS_ntohl( pxTCPHeader->ulSequenceNumber );
-        BaseType_t xSendLength = 0;
-        UBaseType_t uxIntermediateResult = 0U;
-
-        /* Either expect a ACK or a SYN+ACK. */
-        uint8_t ucExpect = tcpTCP_FLAG_ACK;
-        const uint8_t ucFlagsMask = tcpTCP_FLAG_ACK | tcpTCP_FLAG_RST | tcpTCP_FLAG_SYN | tcpTCP_FLAG_FIN;
-
-        if( pxSocket->u.xTCP.ucTCPState == ( uint8_t ) eCONNECT_SYN )
-        {
-            ucExpect |= tcpTCP_FLAG_SYN;
-        }
-
-        if( ( ucTCPFlags & ucFlagsMask ) != ucExpect )
-        {
-            /* eSYN_RECEIVED: flags 0010 expected, not 0002. */
-            /* eSYN_RECEIVED: flags ACK  expected, not SYN. */
-            FreeRTOS_debug_printf( ( "%s: flags %04X expected, not %04X\n",
-                                     ( pxSocket->u.xTCP.ucTCPState == ( uint8_t ) eSYN_RECEIVED ) ? "eSYN_RECEIVED" : "eCONNECT_SYN",
-                                     ucExpect, ucTCPFlags ) );
-
-            /* In case pxSocket is not yet owned by the application, a closure
-             * of the socket will be scheduled for the next cycle. */
-            vTCPStateChange( pxSocket, eCLOSE_WAIT );
-
-            /* Send RST with the expected sequence and ACK numbers,
-             * otherwise the packet will be ignored. */
-            pxTCPWindow->ulOurSequenceNumber = FreeRTOS_htonl( pxTCPHeader->ulAckNr );
-            pxTCPWindow->rx.ulCurrentSequenceNumber = ulSequenceNumber;
-
-            pxTCPHeader->ucTCPFlags |= tcpTCP_FLAG_RST;
-
-            uxIntermediateResult = uxIPHeaderSizeSocket( pxSocket ) + ipSIZE_OF_TCP_HEADER + uxOptionsLength;
-            xSendLength = ( BaseType_t ) uxIntermediateResult;
-
-            pxTCPHeader->ucTCPOffset = ( uint8_t ) ( ( ipSIZE_OF_TCP_HEADER + uxOptionsLength ) << 2 );
-        }
-        else
-        {
-            pxTCPWindow->usPeerPortNumber = pxSocket->u.xTCP.usRemotePort;
-            pxTCPWindow->usOurPortNumber = pxSocket->usLocalPort;
-
-            if( pxSocket->u.xTCP.ucTCPState == ( uint8_t ) eCONNECT_SYN )
-            {
-                /* Map the Last packet onto the ProtocolHeader_t struct for easy access to the fields. */
-                ProtocolHeaders_t * pxLastHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                            &( pxSocket->u.xTCP.xPacket.u.ucLastPacket[ ipSIZE_OF_ETH_HEADER + uxIPHeaderSizeSocket( pxSocket ) ] ) );
-
-                /* Clear the SYN flag in lastPacket. */
-                pxLastHeaders->xTCPHeader.ucTCPFlags = tcpTCP_FLAG_ACK;
-                pxProtocolHeaders->xTCPHeader.ucTCPFlags = tcpTCP_FLAG_ACK;
-
-                /* This socket was the one connecting actively so now perform the
-                 * synchronisation. */
-                vTCPWindowInit( &pxSocket->u.xTCP.xTCPWindow,
-                                ulSequenceNumber, pxSocket->u.xTCP.xTCPWindow.ulOurSequenceNumber, ( uint32_t ) pxSocket->u.xTCP.usMSS );
-                pxTCPWindow->rx.ulHighestSequenceNumber = ulSequenceNumber + 1U;
-                pxTCPWindow->rx.ulCurrentSequenceNumber = ulSequenceNumber + 1U;
-                pxTCPWindow->tx.ulCurrentSequenceNumber++; /* because we send a TCP_SYN [ | TCP_ACK ]; */
-                pxTCPWindow->ulNextTxSequenceNumber++;
-            }
-            else if( ulReceiveLength == 0U )
-            {
-                pxTCPWindow->rx.ulCurrentSequenceNumber = ulSequenceNumber;
-            }
-            else
-            {
-                /* Nothing. */
-            }
-
-            /* The SYN+ACK has been confirmed, increase the next sequence number by
-             * 1. */
-            pxTCPWindow->ulOurSequenceNumber = pxTCPWindow->tx.ulFirstSequenceNumber + 1U;
-
-            #if ( ipconfigUSE_TCP_WIN == 1 )
-                {
-                    FreeRTOS_debug_printf( ( "TCP: %s %u => %xip:%u set ESTAB (scaling %u)\n",
-                                             ( pxSocket->u.xTCP.ucTCPState == ( uint8_t ) eCONNECT_SYN ) ? "active" : "passive",
-                                             pxSocket->usLocalPort,
-                                             ( unsigned ) pxSocket->u.xTCP.ulRemoteIP,
-                                             pxSocket->u.xTCP.usRemotePort,
-                                             ( unsigned ) pxSocket->u.xTCP.bits.bWinScaling ) );
-                }
-            #endif /* ipconfigUSE_TCP_WIN */
-
-            if( ( pxSocket->u.xTCP.ucTCPState == ( EventBits_t ) eCONNECT_SYN ) || ( ulReceiveLength != 0U ) )
-            {
-                pxTCPHeader->ucTCPFlags = tcpTCP_FLAG_ACK;
-
-                uxIntermediateResult = uxIPHeaderSizeSocket( pxSocket ) + ( size_t ) ipSIZE_OF_TCP_HEADER + uxOptionsLength;
-                xSendLength = ( BaseType_t ) uxIntermediateResult;
-                pxTCPHeader->ucTCPOffset = ( uint8_t ) ( ( ipSIZE_OF_TCP_HEADER + uxOptionsLength ) << 2 );
-            }
-
-            #if ( ipconfigUSE_TCP_WIN != 0 )
-                {
-                    if( pxSocket->u.xTCP.bits.bWinScaling == pdFALSE_UNSIGNED )
-                    {
-                        /* The other party did not send a scaling factor.
-                         * A shifting factor in this side must be canceled. */
-                        pxSocket->u.xTCP.ucMyWinScaleFactor = 0;
-                        pxSocket->u.xTCP.ucPeerWinScaleFactor = 0;
-                    }
-                }
-            #endif /* ipconfigUSE_TCP_WIN */
-
-            /* This was the third step of connecting: SYN, SYN+ACK, ACK so now the
-             * connection is established. */
-            vTCPStateChange( pxSocket, eESTABLISHED );
-        }
-
-        return xSendLength;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief prvHandleEstablished(): called from prvTCPHandleState()
- *        Called if the status is eESTABLISHED. Data reception has been handled
- *        earlier. Here the ACK's from peer will be checked, and if a FIN is received,
- *        the code will check if it may be accepted, i.e. if all expected data has been
- *        completely received.
- *
- * @param[in] pxSocket: The socket owning the connection.
- * @param[in,out] ppxNetworkBuffer: Pointer to pointer to the network buffer.
- * @param[in] ulReceiveLength: The length of the received packet.
- * @param[in] uxOptionsLength: Length of TCP options.
- *
- * @return The send length of the packet to be sent.
- */
-    static BaseType_t prvHandleEstablished( FreeRTOS_Socket_t * pxSocket,
-                                            NetworkBufferDescriptor_t ** ppxNetworkBuffer,
-                                            uint32_t ulReceiveLength,
-                                            UBaseType_t uxOptionsLength )
-    {
-        /* Map the buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        ProtocolHeaders_t * pxProtocolHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                        &( ( *ppxNetworkBuffer )->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + uxIPHeaderSizeSocket( pxSocket ) ] ) );
-        TCPHeader_t * pxTCPHeader = &pxProtocolHeaders->xTCPHeader;
-        TCPWindow_t * pxTCPWindow = &pxSocket->u.xTCP.xTCPWindow;
-        uint8_t ucTCPFlags = pxTCPHeader->ucTCPFlags;
-        uint32_t ulSequenceNumber = FreeRTOS_ntohl( pxTCPHeader->ulSequenceNumber ), ulCount, ulIntermediateResult = 0;
-        BaseType_t xSendLength = 0, xMayClose = pdFALSE, bRxComplete, bTxDone;
-        int32_t lDistance, lSendResult;
-        uint16_t usWindow;
-        UBaseType_t uxIntermediateResult = 0;
-
-        /* Remember the window size the peer is advertising. */
-        usWindow = FreeRTOS_ntohs( pxTCPHeader->usWindow );
-        pxSocket->u.xTCP.ulWindowSize = ( uint32_t ) usWindow;
-        #if ( ipconfigUSE_TCP_WIN != 0 )
-            {
-                pxSocket->u.xTCP.ulWindowSize =
-                    ( pxSocket->u.xTCP.ulWindowSize << pxSocket->u.xTCP.ucPeerWinScaleFactor );
-            }
-        #endif /* ipconfigUSE_TCP_WIN */
-
-        if( ( ucTCPFlags & ( uint8_t ) tcpTCP_FLAG_ACK ) != 0U )
-        {
-            ulCount = ulTCPWindowTxAck( pxTCPWindow, FreeRTOS_ntohl( pxTCPHeader->ulAckNr ) );
-
-            /* ulTCPWindowTxAck() returns the number of bytes which have been acked,
-             * starting at 'tx.ulCurrentSequenceNumber'.  Advance the tail pointer in
-             * txStream. */
-            if( ( pxSocket->u.xTCP.txStream != NULL ) && ( ulCount > 0U ) )
-            {
-                /* Just advancing the tail index, 'ulCount' bytes have been
-                 * confirmed, and because there is new space in the txStream, the
-                 * user/owner should be woken up. */
-                /* _HT_ : only in case the socket's waiting? */
-                if( uxStreamBufferGet( pxSocket->u.xTCP.txStream, 0U, NULL, ( size_t ) ulCount, pdFALSE ) != 0U )
-                {
-                    pxSocket->xEventBits |= ( EventBits_t ) eSOCKET_SEND;
-
-                    #if ipconfigSUPPORT_SELECT_FUNCTION == 1
-                        {
-                            if( ( pxSocket->xSelectBits & ( ( EventBits_t ) eSELECT_WRITE ) ) != 0U )
-                            {
-                                pxSocket->xEventBits |= ( ( EventBits_t ) eSELECT_WRITE ) << SOCKET_EVENT_BIT_COUNT;
-                            }
-                        }
-                    #endif
-
-                    /* In case the socket owner has installed an OnSent handler,
-                     * call it now. */
-                    #if ( ipconfigUSE_CALLBACKS == 1 )
-                        {
-                            if( ipconfigIS_VALID_PROG_ADDRESS( pxSocket->u.xTCP.pxHandleSent ) )
-                            {
-                                pxSocket->u.xTCP.pxHandleSent( ( Socket_t ) pxSocket, ulCount );
-                            }
-                        }
-                    #endif /* ipconfigUSE_CALLBACKS == 1  */
-                }
-            }
-        }
-
-        /* If this socket has a stream for transmission, add the data to the
-         * outgoing segment(s). */
-        if( pxSocket->u.xTCP.txStream != NULL )
-        {
-            prvTCPAddTxData( pxSocket );
-        }
-
-        pxSocket->u.xTCP.xTCPWindow.ulOurSequenceNumber = pxTCPWindow->tx.ulCurrentSequenceNumber;
-
-        if( ( pxSocket->u.xTCP.bits.bFinAccepted != pdFALSE_UNSIGNED ) || ( ( ucTCPFlags & ( uint8_t ) tcpTCP_FLAG_FIN ) != 0U ) )
-        {
-            /* Peer is requesting to stop, see if we're really finished. */
-            xMayClose = pdTRUE;
-
-            /* Checks are only necessary if we haven't sent a FIN yet. */
-            if( pxSocket->u.xTCP.bits.bFinSent == pdFALSE_UNSIGNED )
-            {
-                /* xTCPWindowTxDone returns true when all Tx queues are empty. */
-                bRxComplete = xTCPWindowRxEmpty( pxTCPWindow );
-                bTxDone = xTCPWindowTxDone( pxTCPWindow );
-
-                if( ( bRxComplete == 0 ) || ( bTxDone == 0 ) )
-                {
-                    /* Refusing FIN: Rx incomplete 1 optlen 4 tx done 1. */
-                    FreeRTOS_debug_printf( ( "Refusing FIN[%u,%u]: RxCompl %d tx done %d\n",
-                                             pxSocket->usLocalPort,
-                                             pxSocket->u.xTCP.usRemotePort,
-                                             ( int ) bRxComplete,
-                                             ( int ) bTxDone ) );
-                    xMayClose = pdFALSE;
-                }
-                else
-                {
-                    ulIntermediateResult = ulSequenceNumber + ulReceiveLength - pxTCPWindow->rx.ulCurrentSequenceNumber;
-                    lDistance = ( int32_t ) ulIntermediateResult;
-
-                    if( lDistance > 1 )
-                    {
-                        FreeRTOS_debug_printf( ( "Refusing FIN: Rx not complete %d (cur %u high %u)\n",
-                                                 ( int ) lDistance,
-                                                 ( unsigned ) ( pxTCPWindow->rx.ulCurrentSequenceNumber - pxTCPWindow->rx.ulFirstSequenceNumber ),
-                                                 ( unsigned ) ( pxTCPWindow->rx.ulHighestSequenceNumber - pxTCPWindow->rx.ulFirstSequenceNumber ) ) );
-
-                        xMayClose = pdFALSE;
-                    }
-                }
-            }
-
-            if( xTCPWindowLoggingLevel > 0 )
-            {
-                FreeRTOS_debug_printf( ( "TCP: FIN received, mayClose = %d (Rx %u Len %d, Tx %u)\n",
-                                         ( int ) xMayClose,
-                                         ( unsigned ) ( ulSequenceNumber - pxSocket->u.xTCP.xTCPWindow.rx.ulFirstSequenceNumber ),
-                                         ( unsigned ) ulReceiveLength,
-                                         ( unsigned ) ( pxTCPWindow->tx.ulCurrentSequenceNumber - pxSocket->u.xTCP.xTCPWindow.tx.ulFirstSequenceNumber ) ) );
-            }
-
-            if( xMayClose != pdFALSE )
-            {
-                pxSocket->u.xTCP.bits.bFinAccepted = pdTRUE_UNSIGNED;
-                xSendLength = prvTCPHandleFin( pxSocket, *ppxNetworkBuffer );
-            }
-        }
-
-        if( xMayClose == pdFALSE )
-        {
-            pxTCPHeader->ucTCPFlags = tcpTCP_FLAG_ACK;
-
-            if( ulReceiveLength != 0U )
-            {
-                uxIntermediateResult = uxIPHeaderSizeSocket( pxSocket ) + ipSIZE_OF_TCP_HEADER + uxOptionsLength;
-                xSendLength = ( BaseType_t ) uxIntermediateResult;
-                /* TCP-offset equals '( ( length / 4 ) << 4 )', resulting in a shift-left 2 */
-                pxTCPHeader->ucTCPOffset = ( uint8_t ) ( ( ipSIZE_OF_TCP_HEADER + uxOptionsLength ) << 2 );
-
-                if( pxSocket->u.xTCP.bits.bFinSent != pdFALSE_UNSIGNED )
-                {
-                    pxTCPWindow->tx.ulCurrentSequenceNumber = pxTCPWindow->tx.ulFINSequenceNumber;
-                }
-            }
-
-            /* Now get data to be transmitted. */
-
-            /* _HT_ patch: since the MTU has be fixed at 1500 in stead of 1526, TCP
-             * can not send-out both TCP options and also a full packet. Sending
-             * options (SACK) is always more urgent than sending data, which can be
-             * sent later. */
-            if( uxOptionsLength == 0U )
-            {
-                /* prvTCPPrepareSend might allocate a bigger network buffer, if
-                 * necessary. */
-                lSendResult = prvTCPPrepareSend( pxSocket, ppxNetworkBuffer, uxOptionsLength );
-
-                if( lSendResult > 0 )
-                {
-                    xSendLength = ( BaseType_t ) lSendResult;
-                }
-            }
-        }
-
-        return xSendLength;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief Called from prvTCPHandleState(). There is data to be sent. If
- *        ipconfigUSE_TCP_WIN is defined, and if only an ACK must be sent, it will be
- *        checked if it would better be postponed for efficiency.
- *
- * @param[in] pxSocket: The socket owning the TCP connection.
- * @param[in] ppxNetworkBuffer: Pointer to pointer to the network buffer.
- * @param[in] ulReceiveLength: The length of the received buffer.
- * @param[in] xByteCount: Length of the data to be sent.
- *
- * @return The number of bytes actually sent.
- */
-    static BaseType_t prvSendData( FreeRTOS_Socket_t * pxSocket,
-                                   NetworkBufferDescriptor_t ** ppxNetworkBuffer,
-                                   uint32_t ulReceiveLength,
-                                   BaseType_t xByteCount )
-    {
-        /* Map the buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        const ProtocolHeaders_t * pxProtocolHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                              &( ( *ppxNetworkBuffer )->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + xIPHeaderSize( *ppxNetworkBuffer ) ] ) );
-        const TCPHeader_t * pxTCPHeader = &pxProtocolHeaders->xTCPHeader;
-        const TCPWindow_t * pxTCPWindow = &pxSocket->u.xTCP.xTCPWindow;
-        /* Find out what window size we may advertised. */
-        int32_t lRxSpace;
-        BaseType_t xSendLength = xByteCount;
-        uint32_t ulRxBufferSpace;
-        /* Two steps to please MISRA. */
-        size_t uxSize = ipSIZE_OF_IPv4_HEADER + ipSIZE_OF_TCP_HEADER;
-        BaseType_t xSizeWithoutData = ( BaseType_t ) uxSize;
-
-        #if ( ipconfigUSE_TCP_WIN == 1 )
-            int32_t lMinLength;
-        #endif
-
-        /* Set the time-out field, so that we'll be called by the IP-task in case no
-         * next message will be received. */
-        ulRxBufferSpace = pxSocket->u.xTCP.ulHighestRxAllowed - pxTCPWindow->rx.ulCurrentSequenceNumber;
-        lRxSpace = ( int32_t ) ulRxBufferSpace;
-
-        #if ipconfigUSE_TCP_WIN == 1
-            {
-                /* An ACK may be delayed if the peer has space for at least 2 x MSS. */
-                lMinLength = ( ( int32_t ) 2 ) * ( ( int32_t ) pxSocket->u.xTCP.usMSS );
-
-                /* In case we're receiving data continuously, we might postpone sending
-                 * an ACK to gain performance. */
-                /* lint e9007 is OK because 'uxIPHeaderSizeSocket()' has no side-effects. */
-                if( ( ulReceiveLength > 0U ) &&                                    /* Data was sent to this socket. */
-                    ( lRxSpace >= lMinLength ) &&                                  /* There is Rx space for more data. */
-                    ( pxSocket->u.xTCP.bits.bFinSent == pdFALSE_UNSIGNED ) &&      /* Not in a closure phase. */
-                    ( xSendLength == xSizeWithoutData ) &&                         /* No Tx data or options to be sent. */
-                    ( pxSocket->u.xTCP.ucTCPState == ( uint8_t ) eESTABLISHED ) && /* Connection established. */
-                    ( pxTCPHeader->ucTCPFlags == tcpTCP_FLAG_ACK ) )               /* There are no other flags than an ACK. */
-                {
-                    uint32_t ulCurMSS = ( uint32_t ) pxSocket->u.xTCP.usMSS;
-                    int32_t lCurMSS = ( int32_t ) ulCurMSS;
-
-                    if( pxSocket->u.xTCP.pxAckMessage != *ppxNetworkBuffer )
-                    {
-                        /* There was still a delayed in queue, delete it. */
-                        if( pxSocket->u.xTCP.pxAckMessage != NULL )
-                        {
-                            vReleaseNetworkBufferAndDescriptor( pxSocket->u.xTCP.pxAckMessage );
-                        }
-
-                        pxSocket->u.xTCP.pxAckMessage = *ppxNetworkBuffer;
-                    }
-
-                    if( ( ulReceiveLength < ulCurMSS ) || /* Received a small message. */
-                        ( lRxSpace < 2 * lCurMSS ) )      /* There are less than 2 x MSS space in the Rx buffer. */
-                    {
-                        pxSocket->u.xTCP.usTimeout = ( uint16_t ) tcpDELAYED_ACK_SHORT_DELAY_MS;
-                    }
-                    else
-                    {
-                        /* Normally a delayed ACK should wait 200 ms for a next incoming
-                         * packet.  Only wait 20 ms here to gain performance.  A slow ACK
-                         * for full-size message. */
-                        pxSocket->u.xTCP.usTimeout = ( uint16_t ) pdMS_TO_TICKS( tcpDELAYED_ACK_LONGER_DELAY_MS );
-
-                        if( pxSocket->u.xTCP.usTimeout < 1U )
-                        {
-                            pxSocket->u.xTCP.usTimeout = 1U;
-                        }
-                    }
-
-                    if( ( xTCPWindowLoggingLevel > 1 ) && ( ipconfigTCP_MAY_LOG_PORT( pxSocket->usLocalPort ) ) )
-                    {
-                        FreeRTOS_debug_printf( ( "Send[%u->%u] del ACK %u SEQ %u (len %u) tmout %u d %d\n",
-                                                 pxSocket->usLocalPort,
-                                                 pxSocket->u.xTCP.usRemotePort,
-                                                 ( unsigned ) ( pxTCPWindow->rx.ulCurrentSequenceNumber - pxTCPWindow->rx.ulFirstSequenceNumber ),
-                                                 ( unsigned ) ( pxSocket->u.xTCP.xTCPWindow.ulOurSequenceNumber - pxTCPWindow->tx.ulFirstSequenceNumber ),
-                                                 ( unsigned ) xSendLength,
-                                                 pxSocket->u.xTCP.usTimeout,
-                                                 ( int ) lRxSpace ) );
-                    }
-
-                    *ppxNetworkBuffer = NULL;
-                    xSendLength = 0;
-                }
-                else if( pxSocket->u.xTCP.pxAckMessage != NULL )
-                {
-                    /* As an ACK is not being delayed, remove any earlier delayed ACK
-                     * message. */
-                    if( pxSocket->u.xTCP.pxAckMessage != *ppxNetworkBuffer )
-                    {
-                        vReleaseNetworkBufferAndDescriptor( pxSocket->u.xTCP.pxAckMessage );
-                    }
-
-                    pxSocket->u.xTCP.pxAckMessage = NULL;
-                }
-                else
-                {
-                    /* The ack will not be postponed, and there was no stored ack ( in 'pxAckMessage' ). */
-                }
-            }
-        #else /* if ipconfigUSE_TCP_WIN == 1 */
-            {
-                /* Remove compiler warnings. */
-                ( void ) ulReceiveLength;
-                ( void ) pxTCPHeader;
-                ( void ) lRxSpace;
-            }
-        #endif /* ipconfigUSE_TCP_WIN */
-
-        if( xSendLength != 0 )
-        {
-            if( ( xTCPWindowLoggingLevel > 1 ) && ( ipconfigTCP_MAY_LOG_PORT( pxSocket->usLocalPort ) ) )
-            {
-                FreeRTOS_debug_printf( ( "Send[%u->%u] imm ACK %u SEQ %u (len %u)\n",
-                                         pxSocket->usLocalPort,
-                                         pxSocket->u.xTCP.usRemotePort,
-                                         ( unsigned ) ( pxTCPWindow->rx.ulCurrentSequenceNumber - pxTCPWindow->rx.ulFirstSequenceNumber ),
-                                         ( unsigned ) ( pxTCPWindow->ulOurSequenceNumber - pxTCPWindow->tx.ulFirstSequenceNumber ),
-                                         ( unsigned ) xSendLength ) );
-            }
-
-            /* Set the parameter 'xReleaseAfterSend' to the value of
-             * ipconfigZERO_COPY_TX_DRIVER. */
-            prvTCPReturnPacket( pxSocket, *ppxNetworkBuffer, ( uint32_t ) xSendLength, ipconfigZERO_COPY_TX_DRIVER );
-            #if ( ipconfigZERO_COPY_TX_DRIVER != 0 )
-                {
-                    /* The driver has taken ownership of the Network Buffer. */
-                    *ppxNetworkBuffer = NULL;
-                }
-            #endif
-        }
-
-        return xSendLength;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief Check incoming packets for valid data and handle the state of the
- *        TCP connection and respond according to the situation.
- *
- * @param[in] pxSocket: The socket whose connection state is being handled.
- * @param[in] ppxNetworkBuffer: The network buffer descriptor holding the
- *            packet received from the peer.
- *
- * @return If the data is correct and some packet was sent to the peer, then
- *         the number of bytes sent is returned, or else a negative value is
- *         returned indicating an error.
- *
- * @note prvTCPHandleState() is the most important function of this TCP stack
- * We've tried to keep it (relatively short) by putting a lot of code in
- * the static functions above:
- *
- *      prvCheckRxData()
- *      prvStoreRxData()
- *      prvSetOptions()
- *      prvHandleSynReceived()
- *      prvHandleEstablished()
- *      prvSendData()
- *
- * As these functions are declared static, and they're called from one location
- * only, most compilers will inline them, thus avoiding a call and return.
- */
-    static BaseType_t prvTCPHandleState( FreeRTOS_Socket_t * pxSocket,
-                                         NetworkBufferDescriptor_t ** ppxNetworkBuffer )
-    {
-        /* Map the buffer onto the ProtocolHeader_t struct for easy access to the fields. */
-        ProtocolHeaders_t * pxProtocolHeaders = ipCAST_PTR_TO_TYPE_PTR( ProtocolHeaders_t,
-                                                                        &( ( *ppxNetworkBuffer )->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + xIPHeaderSize( *ppxNetworkBuffer ) ] ) );
-        TCPHeader_t * pxTCPHeader = &( pxProtocolHeaders->xTCPHeader );
-        BaseType_t xSendLength = 0;
-        uint32_t ulReceiveLength; /* Number of bytes contained in the TCP message. */
-        uint8_t * pucRecvData;
-        uint32_t ulSequenceNumber = FreeRTOS_ntohl( pxTCPHeader->ulSequenceNumber );
-
-        /* uxOptionsLength: the size of the options to be sent (always a multiple of
-         * 4 bytes)
-         * 1. in the SYN phase, we shall communicate the MSS
-         * 2. in case of a SACK, Selective ACK, ack a segment which comes in
-         * out-of-order. */
-        UBaseType_t uxOptionsLength = 0U;
-        uint8_t ucTCPFlags = pxTCPHeader->ucTCPFlags;
-        TCPWindow_t * pxTCPWindow = &( pxSocket->u.xTCP.xTCPWindow );
-        UBaseType_t uxIntermediateResult = 0;
-        uint32_t ulSum;
-
-        /* First get the length and the position of the received data, if any.
-         * pucRecvData will point to the first byte of the TCP payload. */
-        ulReceiveLength = ( uint32_t ) prvCheckRxData( *ppxNetworkBuffer, &pucRecvData );
-
-        if( pxSocket->u.xTCP.ucTCPState >= ( uint8_t ) eESTABLISHED )
-        {
-            if( pxTCPWindow->rx.ulCurrentSequenceNumber == ( ulSequenceNumber + 1U ) )
-            {
-                /* This is most probably a keep-alive message from peer.  Setting
-                 * 'bWinChange' doesn't cause a window-size-change, the flag is used
-                 * here to force sending an immediate ACK. */
-                pxSocket->u.xTCP.bits.bWinChange = pdTRUE_UNSIGNED;
-            }
-        }
-
-        /* Keep track of the highest sequence number that might be expected within
-         * this connection. */
-        ulSum = ulSequenceNumber + ulReceiveLength - pxTCPWindow->rx.ulHighestSequenceNumber;
-
-        if( ( ( int32_t ) ulSum ) > 0 )
-        {
-            pxTCPWindow->rx.ulHighestSequenceNumber = ulSequenceNumber + ulReceiveLength;
-        }
-
-        /* Storing data may result in a fatal error if malloc() fails. */
-        if( prvStoreRxData( pxSocket, pucRecvData, *ppxNetworkBuffer, ulReceiveLength ) < 0 )
-        {
-            xSendLength = -1;
-        }
-        else
-        {
-            eIPTCPState_t eState;
-
-            uxOptionsLength = prvSetOptions( pxSocket, *ppxNetworkBuffer );
-
-            if( ( pxSocket->u.xTCP.ucTCPState == ( uint8_t ) eSYN_RECEIVED ) && ( ( ucTCPFlags & ( uint8_t ) tcpTCP_FLAG_CTRL ) == ( uint8_t ) tcpTCP_FLAG_SYN ) )
-            {
-                FreeRTOS_debug_printf( ( "eSYN_RECEIVED: ACK expected, not SYN: peer missed our SYN+ACK\n" ) );
-
-                /* In eSYN_RECEIVED a simple ACK is expected, but apparently the
-                 * 'SYN+ACK' didn't arrive.  Step back to the previous state in which
-                 * a first incoming SYN is handled.  The SYN was counted already so
-                 * decrease it first. */
-                vTCPStateChange( pxSocket, eSYN_FIRST );
-            }
-
-            if( ( ( ucTCPFlags & tcpTCP_FLAG_FIN ) != 0U ) && ( pxSocket->u.xTCP.bits.bFinRecv == pdFALSE_UNSIGNED ) )
-            {
-                /* It's the first time a FIN has been received, remember its
-                 * sequence number. */
-                pxTCPWindow->rx.ulFINSequenceNumber = ulSequenceNumber + ulReceiveLength;
-                pxSocket->u.xTCP.bits.bFinRecv = pdTRUE_UNSIGNED;
-
-                /* Was peer the first one to send a FIN? */
-                if( pxSocket->u.xTCP.bits.bFinSent == pdFALSE_UNSIGNED )
-                {
-                    /* If so, don't send the-last-ACK. */
-                    pxSocket->u.xTCP.bits.bFinLast = pdTRUE_UNSIGNED;
-                }
-            }
-
-            eState = pxSocket->u.xTCP.ucTCPState;
-
-            switch( eState )
-            {
-                case eCLOSED: /* (server + client) no connection state at all. */
-
-                    /* Nothing to do for a closed socket, except waiting for the
-                     * owner. */
-                    break;
-
-                case eTCP_LISTEN: /* (server) waiting for a connection request from
-                                   * any remote TCP and port. */
-
-                    /* The listen state was handled in xProcessReceivedTCPPacket().
-                     * Should not come here. */
-                    break;
-
-                case eSYN_FIRST: /* (server) Just received a SYN request for a server
-                                  * socket. */
-
-                    /* A new socket has been created, reply with a SYN+ACK.
-                     * Acknowledge with seq+1 because the SYN is seen as pseudo data
-                     * with len = 1. */
-                    uxOptionsLength = prvSetSynAckOptions( pxSocket, pxTCPHeader );
-                    pxTCPHeader->ucTCPFlags = ( uint8_t ) tcpTCP_FLAG_SYN | ( uint8_t ) tcpTCP_FLAG_ACK;
-
-                    uxIntermediateResult = uxIPHeaderSizeSocket( pxSocket ) + ipSIZE_OF_TCP_HEADER + uxOptionsLength;
-                    xSendLength = ( BaseType_t ) uxIntermediateResult;
-
-                    /* Set the TCP offset field:  ipSIZE_OF_TCP_HEADER equals 20 and
-                     * uxOptionsLength is a multiple of 4.  The complete expression is:
-                     * ucTCPOffset = ( ( ipSIZE_OF_TCP_HEADER + uxOptionsLength ) / 4 ) << 4 */
-                    pxTCPHeader->ucTCPOffset = ( uint8_t ) ( ( ipSIZE_OF_TCP_HEADER + uxOptionsLength ) << 2 );
-                    vTCPStateChange( pxSocket, eSYN_RECEIVED );
-
-                    pxTCPWindow->rx.ulHighestSequenceNumber = ulSequenceNumber + 1U;
-                    pxTCPWindow->rx.ulCurrentSequenceNumber = ulSequenceNumber + 1U;
-                    pxTCPWindow->ulNextTxSequenceNumber = pxTCPWindow->tx.ulFirstSequenceNumber + 1U;
-                    pxTCPWindow->tx.ulCurrentSequenceNumber = pxTCPWindow->tx.ulFirstSequenceNumber + 1U; /* because we send a TCP_SYN. */
-                    break;
-
-                case eCONNECT_SYN:  /* (client) also called SYN_SENT: we've just send a
-                                     * SYN, expect a SYN+ACK and send a ACK now. */
-                /* Fall through */
-                case eSYN_RECEIVED: /* (server) we've had a SYN, replied with SYN+SCK
-                                     * expect a ACK and do nothing. */
-                    xSendLength = prvHandleSynReceived( pxSocket, *( ppxNetworkBuffer ), ulReceiveLength, uxOptionsLength );
-                    break;
-
-                case eESTABLISHED: /* (server + client) an open connection, data
-                                    * received can be delivered to the user. The normal
-                                    * state for the data transfer phase of the connection
-                                    * The closing states are also handled here with the
-                                    * use of some flags. */
-                    xSendLength = prvHandleEstablished( pxSocket, ppxNetworkBuffer, ulReceiveLength, uxOptionsLength );
-                    break;
-
-                case eLAST_ACK:   /* (server + client) waiting for an acknowledgement
-                                   * of the connection termination request previously
-                                   * sent to the remote TCP (which includes an
-                                   * acknowledgement of its connection termination
-                                   * request). */
-                /* Fall through */
-                case eFIN_WAIT_1: /* (server + client) waiting for a connection termination request from the remote TCP,
-                                   * or an acknowledgement of the connection termination request previously sent. */
-                /* Fall through */
-                case eFIN_WAIT_2: /* (server + client) waiting for a connection termination request from the remote TCP. */
-                    xSendLength = prvTCPHandleFin( pxSocket, *ppxNetworkBuffer );
-                    break;
-
-                case eCLOSE_WAIT: /* (server + client) waiting for a connection
-                                   * termination request from the local user.  Nothing to
-                                   * do, connection is closed, wait for owner to close
-                                   * this socket. */
-                    break;
-
-                case eCLOSING: /* (server + client) waiting for a connection
-                                * termination request acknowledgement from the remote
-                                * TCP. */
-                    break;
-
-                case eTIME_WAIT: /* (either server or client) waiting for enough time
-                                  * to pass to be sure the remote TCP received the
-                                  * acknowledgement of its connection termination
-                                  * request. [According to RFC 793 a connection can stay
-                                  * in TIME-WAIT for a maximum of four minutes known as
-                                  * a MSL (maximum segment lifetime).]  These states are
-                                  * implemented implicitly by settings flags like
-                                  * 'bFinSent', 'bFinRecv', and 'bFinAcked'. */
-                    break;
-
-                default:
-                    /* No more known states. */
-                    break;
-            }
-        }
-
-        if( xSendLength > 0 )
-        {
-            xSendLength = prvSendData( pxSocket, ppxNetworkBuffer, ulReceiveLength, xSendLength );
-        }
-
-        return xSendLength;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief Common code for sending a TCP protocol control packet (i.e. no options, no
- *        payload, just flags).
- *
- * @param[in] pxNetworkBuffer: The network buffer received from the peer.
- * @param[in] ucTCPFlags: The flags to determine what kind of packet this is.
- *
- * @return pdFAIL always indicating that the packet was not consumed.
- */
-    static BaseType_t prvTCPSendSpecialPacketHelper( NetworkBufferDescriptor_t * pxNetworkBuffer,
-                                                     uint8_t ucTCPFlags )
-    {
-        #if ( ipconfigIGNORE_UNKNOWN_PACKETS == 1 )
-            /* Configured to ignore unknown packets just suppress a compiler warning. */
-            ( void ) pxNetworkBuffer;
-            ( void ) ucTCPFlags;
-        #else
-            {
-                /* Map the ethernet buffer onto the TCPPacket_t struct for easy access to the fields. */
-                TCPPacket_t * pxTCPPacket = ipCAST_PTR_TO_TYPE_PTR( TCPPacket_t, pxNetworkBuffer->pucEthernetBuffer );
-                const uint32_t ulSendLength =
-                    ( ipSIZE_OF_IPv4_HEADER + ipSIZE_OF_TCP_HEADER ); /* Plus 0 options. */
-
-                pxTCPPacket->xTCPHeader.ucTCPFlags = ucTCPFlags;
-                pxTCPPacket->xTCPHeader.ucTCPOffset = ( ipSIZE_OF_TCP_HEADER ) << 2;
-
-                prvTCPReturnPacket( NULL, pxNetworkBuffer, ulSendLength, pdFALSE );
-            }
-        #endif /* !ipconfigIGNORE_UNKNOWN_PACKETS */
-
-        /* The packet was not consumed. */
-        return pdFAIL;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief A "challenge ACK" is as per https://tools.ietf.org/html/rfc5961#section-3.2,
- *        case #3. In summary, an RST was received with a sequence number that is
- *        unexpected but still within the window.
- *
- * @param[in] pxNetworkBuffer: The network buffer descriptor with the packet.
- *
- * @return Returns the value back from #prvTCPSendSpecialPacketHelper.
- */
-    static BaseType_t prvTCPSendChallengeAck( NetworkBufferDescriptor_t * pxNetworkBuffer )
-    {
-        return prvTCPSendSpecialPacketHelper( pxNetworkBuffer, tcpTCP_FLAG_ACK );
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief Send a RST (Reset) to peer in case the packet cannot be handled.
- *
- * @param[in] pxNetworkBuffer: The network buffer descriptor with the packet.
- *
- * @return Returns the value back from #prvTCPSendSpecialPacketHelper.
- */
-    static BaseType_t prvTCPSendReset( NetworkBufferDescriptor_t * pxNetworkBuffer )
-    {
-        return prvTCPSendSpecialPacketHelper( pxNetworkBuffer,
-                                              ( uint8_t ) tcpTCP_FLAG_ACK | ( uint8_t ) tcpTCP_FLAG_RST );
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief Set the MSS (Maximum segment size) associated with the given socket.
- *
- * @param[in] pxSocket: The socket whose MSS is to be set.
- */
-    static void prvSocketSetMSS( FreeRTOS_Socket_t * pxSocket )
-    {
-        uint32_t ulMSS;
-
-        /* Do not allow MSS smaller than tcpMINIMUM_SEGMENT_LENGTH. */
-        #if ( ipconfigTCP_MSS >= tcpMINIMUM_SEGMENT_LENGTH )
-            {
-                ulMSS = ipconfigTCP_MSS;
-            }
-        #else
-            {
-                ulMSS = tcpMINIMUM_SEGMENT_LENGTH;
-            }
-        #endif
-
-        if( ( ( FreeRTOS_ntohl( pxSocket->u.xTCP.ulRemoteIP ) ^ *ipLOCAL_IP_ADDRESS_POINTER ) & xNetworkAddressing.ulNetMask ) != 0U )
-        {
-            /* Data for this peer will pass through a router, and maybe through
-             * the internet.  Limit the MSS to 1400 bytes or less. */
-            ulMSS = FreeRTOS_min_uint32( ( uint32_t ) tcpREDUCED_MSS_THROUGH_INTERNET, ulMSS );
-        }
-
-        FreeRTOS_debug_printf( ( "prvSocketSetMSS: %u bytes for %xip:%u\n", ( unsigned ) ulMSS, ( unsigned ) pxSocket->u.xTCP.ulRemoteIP, pxSocket->u.xTCP.usRemotePort ) );
-
-        pxSocket->u.xTCP.usMSS = ( uint16_t ) ulMSS;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
  * @brief Process the received TCP packet.
  *
  * @param[in] pxDescriptor: The descriptor in which the TCP packet is held.
@@ -2260,263 +1070,6 @@
     }
     /*-----------------------------------------------------------*/
 
-/**
- * @brief Handle 'listen' event on the given socket.
- *
- * @param[in] pxSocket: The socket on which the listen occurred.
- * @param[in] pxNetworkBuffer: The network buffer carrying the packet.
- *
- * @return If a new socket/duplicate socket is created, then the pointer to
- *         that socket is returned or else, a NULL pointer is returned.
- */
-    static FreeRTOS_Socket_t * prvHandleListen( FreeRTOS_Socket_t * pxSocket,
-                                                NetworkBufferDescriptor_t * pxNetworkBuffer )
-    {
-        /* Map the ethernet buffer onto a TCPPacket_t struct for easy access to the fields. */
-        const TCPPacket_t * pxTCPPacket = ipCAST_CONST_PTR_TO_CONST_TYPE_PTR( TCPPacket_t, pxNetworkBuffer->pucEthernetBuffer );
-        FreeRTOS_Socket_t * pxReturn = NULL;
-        uint32_t ulInitialSequenceNumber;
-
-        /* Silently discard a SYN packet which was not specifically sent for this node. */
-        if( pxTCPPacket->xIPHeader.ulDestinationIPAddress == *ipLOCAL_IP_ADDRESS_POINTER )
-        {
-            /* Assume that a new Initial Sequence Number will be required. Request
-             * it now in order to fail out if necessary. */
-            ulInitialSequenceNumber = ulApplicationGetNextSequenceNumber( *ipLOCAL_IP_ADDRESS_POINTER,
-                                                                          pxSocket->usLocalPort,
-                                                                          pxTCPPacket->xIPHeader.ulSourceIPAddress,
-                                                                          pxTCPPacket->xTCPHeader.usSourcePort );
-        }
-        else
-        {
-            /* Set the sequence number to 0 to avoid further processing. */
-            ulInitialSequenceNumber = 0U;
-        }
-
-        /* A pure SYN (without ACK) has come in, create a new socket to answer
-         * it. */
-        if( ulInitialSequenceNumber != 0U )
-        {
-            if( pxSocket->u.xTCP.bits.bReuseSocket != pdFALSE_UNSIGNED )
-            {
-                /* The flag bReuseSocket indicates that the same instance of the
-                 * listening socket should be used for the connection. */
-                pxReturn = pxSocket;
-                pxSocket->u.xTCP.bits.bPassQueued = pdTRUE_UNSIGNED;
-                pxSocket->u.xTCP.pxPeerSocket = pxSocket;
-            }
-            else
-            {
-                /* The socket does not have the bReuseSocket flag set meaning create a
-                 * new socket when a connection comes in. */
-                pxReturn = NULL;
-
-                if( pxSocket->u.xTCP.usChildCount >= pxSocket->u.xTCP.usBacklog )
-                {
-                    FreeRTOS_printf( ( "Check: Socket %u already has %u / %u child%s\n",
-                                       pxSocket->usLocalPort,
-                                       pxSocket->u.xTCP.usChildCount,
-                                       pxSocket->u.xTCP.usBacklog,
-                                       ( pxSocket->u.xTCP.usChildCount == 1U ) ? "" : "ren" ) );
-                    ( void ) prvTCPSendReset( pxNetworkBuffer );
-                }
-                else
-                {
-                    FreeRTOS_Socket_t * pxNewSocket = ( FreeRTOS_Socket_t * )
-                                                      FreeRTOS_socket( FREERTOS_AF_INET, FREERTOS_SOCK_STREAM, FREERTOS_IPPROTO_TCP );
-
-                    if( ( pxNewSocket == NULL ) || ( pxNewSocket == FREERTOS_INVALID_SOCKET ) )
-                    {
-                        FreeRTOS_debug_printf( ( "TCP: Listen: new socket failed\n" ) );
-                        ( void ) prvTCPSendReset( pxNetworkBuffer );
-                    }
-                    else if( prvTCPSocketCopy( pxNewSocket, pxSocket ) != pdFALSE )
-                    {
-                        /* The socket will be connected immediately, no time for the
-                         * owner to setsockopt's, therefore copy properties of the server
-                         * socket to the new socket.  Only the binding might fail (due to
-                         * lack of resources). */
-                        pxReturn = pxNewSocket;
-                    }
-                    else
-                    {
-                        /* Copying failed somehow. */
-                    }
-                }
-            }
-        }
-
-        if( ( ulInitialSequenceNumber != 0U ) && ( pxReturn != NULL ) )
-        {
-            /* Map the byte stream onto the ProtocolHeaders_t for easy access to the fields. */
-            const ProtocolHeaders_t * pxProtocolHeaders = ipCAST_CONST_PTR_TO_CONST_TYPE_PTR( ProtocolHeaders_t,
-                                                                                              &( pxNetworkBuffer->pucEthernetBuffer[ ipSIZE_OF_ETH_HEADER + xIPHeaderSize( pxNetworkBuffer ) ] ) );
-
-            pxReturn->u.xTCP.usRemotePort = FreeRTOS_htons( pxTCPPacket->xTCPHeader.usSourcePort );
-            pxReturn->u.xTCP.ulRemoteIP = FreeRTOS_htonl( pxTCPPacket->xIPHeader.ulSourceIPAddress );
-            pxReturn->u.xTCP.xTCPWindow.ulOurSequenceNumber = ulInitialSequenceNumber;
-
-            /* Here is the SYN action. */
-            pxReturn->u.xTCP.xTCPWindow.rx.ulCurrentSequenceNumber = FreeRTOS_ntohl( pxProtocolHeaders->xTCPHeader.ulSequenceNumber );
-            prvSocketSetMSS( pxReturn );
-
-            prvTCPCreateWindow( pxReturn );
-
-            vTCPStateChange( pxReturn, eSYN_FIRST );
-
-            /* Make a copy of the header up to the TCP header.  It is needed later
-             * on, whenever data must be sent to the peer. */
-            ( void ) memcpy( ( void * ) pxReturn->u.xTCP.xPacket.u.ucLastPacket,
-                             ( const void * ) pxNetworkBuffer->pucEthernetBuffer,
-                             sizeof( pxReturn->u.xTCP.xPacket.u.ucLastPacket ) );
-        }
-
-        return pxReturn;
-    }
-    /*-----------------------------------------------------------*/
-
-/**
- * @brief Duplicates a socket after a listening socket receives a connection and bind
- *        the new socket to the same port as the listening socket.
- *        Also, let the new socket inherit all properties from the listening socket.
- *
- * @param[in] pxNewSocket: Pointer to the new socket.
- * @param[in] pxSocket: Pointer to the socket being duplicated.
- *
- * @return If all steps all successful, then pdTRUE is returned. Else, pdFALSE.
- */
-    static BaseType_t prvTCPSocketCopy( FreeRTOS_Socket_t * pxNewSocket,
-                                        FreeRTOS_Socket_t * pxSocket )
-    {
-        struct freertos_sockaddr xAddress;
-        BaseType_t xResult;
-
-        pxNewSocket->xReceiveBlockTime = pxSocket->xReceiveBlockTime;
-        pxNewSocket->xSendBlockTime = pxSocket->xSendBlockTime;
-        pxNewSocket->ucSocketOptions = pxSocket->ucSocketOptions;
-        pxNewSocket->u.xTCP.uxRxStreamSize = pxSocket->u.xTCP.uxRxStreamSize;
-        pxNewSocket->u.xTCP.uxTxStreamSize = pxSocket->u.xTCP.uxTxStreamSize;
-        pxNewSocket->u.xTCP.uxLittleSpace = pxSocket->u.xTCP.uxLittleSpace;
-        pxNewSocket->u.xTCP.uxEnoughSpace = pxSocket->u.xTCP.uxEnoughSpace;
-        pxNewSocket->u.xTCP.uxRxWinSize = pxSocket->u.xTCP.uxRxWinSize;
-        pxNewSocket->u.xTCP.uxTxWinSize = pxSocket->u.xTCP.uxTxWinSize;
-
-        #if ( ipconfigSOCKET_HAS_USER_SEMAPHORE == 1 )
-            {
-                pxNewSocket->pxUserSemaphore = pxSocket->pxUserSemaphore;
-            }
-        #endif /* ipconfigSOCKET_HAS_USER_SEMAPHORE */
-
-        #if ( ipconfigUSE_CALLBACKS == 1 )
-            {
-                /* In case call-backs are used, copy them from parent to child. */
-                pxNewSocket->u.xTCP.pxHandleConnected = pxSocket->u.xTCP.pxHandleConnected;
-                pxNewSocket->u.xTCP.pxHandleReceive = pxSocket->u.xTCP.pxHandleReceive;
-                pxNewSocket->u.xTCP.pxHandleSent = pxSocket->u.xTCP.pxHandleSent;
-            }
-        #endif /* ipconfigUSE_CALLBACKS */
-
-        #if ( ipconfigSUPPORT_SELECT_FUNCTION == 1 )
-            {
-                /* Child socket of listening sockets will inherit the Socket Set
-                 * Otherwise the owner has no chance of including it into the set. */
-                if( pxSocket->pxSocketSet != NULL )
-                {
-                    pxNewSocket->pxSocketSet = pxSocket->pxSocketSet;
-                    pxNewSocket->xSelectBits = pxSocket->xSelectBits | ( ( EventBits_t ) eSELECT_READ ) | ( ( EventBits_t ) eSELECT_EXCEPT );
-                }
-            }
-        #endif /* ipconfigSUPPORT_SELECT_FUNCTION */
-
-        /* And bind it to the same local port as its parent. */
-        xAddress.sin_addr = *ipLOCAL_IP_ADDRESS_POINTER;
-        xAddress.sin_port = FreeRTOS_htons( pxSocket->usLocalPort );
-
-        #if ( ipconfigTCP_HANG_PROTECTION == 1 )
-            {
-                /* Only when there is anti-hanging protection, a socket may become an
-                 * orphan temporarily.  Once this socket is really connected, the owner of
-                 * the server socket will be notified. */
-
-                /* When bPassQueued is true, the socket is an orphan until it gets
-                 * connected. */
-                pxNewSocket->u.xTCP.bits.bPassQueued = pdTRUE_UNSIGNED;
-                pxNewSocket->u.xTCP.pxPeerSocket = pxSocket;
-            }
-        #else
-            {
-                /* A reference to the new socket may be stored and the socket is marked
-                 * as 'passable'. */
-
-                /* When bPassAccept is true, this socket may be returned in a call to
-                 * accept(). */
-                pxNewSocket->u.xTCP.bits.bPassAccept = pdTRUE_UNSIGNED;
-
-                if( pxSocket->u.xTCP.pxPeerSocket == NULL )
-                {
-                    pxSocket->u.xTCP.pxPeerSocket = pxNewSocket;
-                }
-            }
-        #endif /* if ( ipconfigTCP_HANG_PROTECTION == 1 ) */
-
-        pxSocket->u.xTCP.usChildCount++;
-
-        FreeRTOS_debug_printf( ( "Gain: Socket %u now has %u / %u child%s\n",
-                                 pxSocket->usLocalPort,
-                                 pxSocket->u.xTCP.usChildCount,
-                                 pxSocket->u.xTCP.usBacklog,
-                                 ( pxSocket->u.xTCP.usChildCount == 1U ) ? "" : "ren" ) );
-
-        /* Now bind the child socket to the same port as the listening socket. */
-        if( vSocketBind( pxNewSocket, &xAddress, sizeof( xAddress ), pdTRUE ) != 0 )
-        {
-            FreeRTOS_debug_printf( ( "TCP: Listen: new socket bind error\n" ) );
-            ( void ) vSocketClose( pxNewSocket );
-            xResult = pdFALSE;
-        }
-        else
-        {
-            xResult = pdTRUE;
-        }
-
-        return xResult;
-    }
-    /*-----------------------------------------------------------*/
-
-    #if ( ( ipconfigHAS_DEBUG_PRINTF != 0 ) || ( ipconfigHAS_PRINTF != 0 ) )
-
-        const char * FreeRTOS_GetTCPStateName( UBaseType_t ulState )
-        {
-            static const char * const pcStateNames[] =
-            {
-                "eCLOSED",
-                "eTCP_LISTEN",
-                "eCONNECT_SYN",
-                "eSYN_FIRST",
-                "eSYN_RECEIVED",
-                "eESTABLISHED",
-                "eFIN_WAIT_1",
-                "eFIN_WAIT_2",
-                "eCLOSE_WAIT",
-                "eCLOSING",
-                "eLAST_ACK",
-                "eTIME_WAIT",
-                "eUNKNOWN",
-            };
-            BaseType_t xIndex = ( BaseType_t ) ulState;
-
-            if( ( xIndex < 0 ) || ( xIndex >= ARRAY_SIZE( pcStateNames ) ) )
-            {
-                /* The last item is called 'eUNKNOWN' */
-                xIndex = ARRAY_SIZE( pcStateNames );
-                xIndex--;
-            }
-
-            return pcStateNames[ xIndex ];
-        }
-
-    #endif /* ( ( ipconfigHAS_DEBUG_PRINTF != 0 ) || ( ipconfigHAS_PRINTF != 0 ) ) */
-    /*-----------------------------------------------------------*/
 
 /**
  * @brief In the API accept(), the user asks is there is a new client? As API's can
